Create endpoints to update, get, and delete boxes

+++ migrations/20231030190410-create-box.js
@@ -13,7 +13,7 @@ module.exports = {
 				type: Sequelize.STRING,
 			},
 			price: {
-				type: Sequelize.FLOAT,
+				type: Sequelize.FLOAT(10, 2),
 			},
 			description: {
 				type: Sequelize.STRING,
